Add (long long) casts.

+++ src/utility/main.c
@@ -295,9 +295,10 @@ static void PrintLogInfo(KineticLogInfo_Type type, KineticLogInfo* info)
         case KINETIC_DEVICE_INFO_TYPE_STATISTICS:
             printf("Device Statistics:\n");
             for (i = 0; i < info->numStatistics; i++) {
-                printf("  %s: count=%lu, bytes=%lu\n",
+                printf("  %s: count=%llu, bytes=%llu\n",
                     KineticMessageType_GetName(info->statistics[i].messageType),
-                    info->statistics[i].count, info->statistics[i].bytes);
+                    (long long)info->statistics[i].count,
+                    (long long)info->statistics[i].bytes);
             }
             break;
